Manually fix react/no-multi-comp violations
