+++ js/components/related-dag.jsx
@@ -1,98 +1,8 @@
 import React from 'react';
-import {
-  COLOUR_MAP,
-  COLOUR_SELECTED,
-  COLOUR_HOVER,
-} from '../utils';
+import DAGEdge from './DAGEdge';
+import DAGNode from './DAGNode';
 
 const {dagre} = window;
-
-/**
- * Renders an edge in the DAG view.
- */
-class DAGEdge extends React.Component {
-  render() {
-    const {points} = this.props;
-    let path = `M ${points[0].x},${points[0].y}`;
-    for (let i = 1; i < points.length; i += 1) {
-      path += ` L ${points[i].x},${points[i].y}`;
-    }
-    return (
-      <path
-        d={path}
-        style={{
-          markerEnd: 'url(#arrow)',
-          stroke: 'black',
-          strokeWidth: 2,
-          fill: 'transparent',
-        }}
-      />
-    );
-  }
-}
-
-/**
- * Renders a node in the DAG view.
- */
-class DAGNode extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.handleOnMouseEnter = this.handleOnMouseEnter.bind(this);
-    this.handleOnMouseLeave = this.handleOnMouseLeave.bind(this);
-
-    this.state = {
-      hover: false,
-    };
-  }
-
-  handleOnMouseEnter() {
-    this.setState({hover: true});
-    if (this.props.hover) {
-      this.props.hover(this.props.node.job);
-    }
-  }
-
-  handleOnMouseLeave() {
-    this.setState({hover: false});
-    if (this.props.hover) {
-      this.props.hover(null);
-    }
-  }
-
-  render() {
-    const {node, selected} = this.props;
-
-    let colour;
-    if (selected) {
-      colour = COLOUR_SELECTED;
-    } else if (this.state.hover) {
-      colour = COLOUR_HOVER;
-    } else {
-      colour = COLOUR_MAP;
-    }
-
-    return (
-      <g
-        onMouseEnter={this.handleOnMouseEnter}
-        onMouseLeave={this.handleOnMouseLeave}
-        transform={`translate(${node.x}, ${node.y})`}
-      >
-        <a href={`/#/job/${node.job.id}`}>
-          <circle r="20" style={{fill: colour}} />
-          <text
-            style={{
-              textAnchor: 'middle',
-              alignmentBaseline: 'middle',
-            }}
-          >
-            {node.label}
-          </text>
-        </a>
-      </g>
-    );
-  }
-}
 
 /**
  * View rendering the DAG of related scalding steps.
