+++ js/job.jsx
@@ -1,25 +1,22 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import {Link} from 'react-router';
 
-import ProgressBar from './components/progress-bar';
+import KillModal from './components/KillModal';
+import MapSummary from './components/MapSummary';
+import ReduceSummary from './components/ReduceSummary';
 import RelatedDAG from './components/related-dag';
+import RelatedJobs from './components/RelatedJobs';
+import TaskStats from './components/TaskStats';
+import TaskWaterfall from './components/TaskWaterfall';
 import {Store} from './store';
 import {
-  lolhadoop,
-  jobState,
   ACTIVE_STATES,
-  FAILED_STATES,
-  timeFormat,
-  secondFormat,
   cleanJobPath,
-  humanFormat,
-  numFormat,
-  plural,
-  COLOUR_MAP,
-  COLOUR_REDUCE,
-  COLOUR_SELECTED,
-  COLOUR_HOVER,
+  FAILED_STATES,
+  jobState,
+  lolhadoop,
+  secondFormat,
+  timeFormat,
 } from './utils';
 import {notAvailable} from './mr';
 
@@ -246,288 +243,5 @@ export default class extends React.Component {
     );
     console.timeEnd('Render Job');
     return rv;
-  }
-}
-
-class KillModal extends React.Component {
-  render() {
-    const {killJob, onHideModal} = this.props;
-    return (
-      <div className="modal show" tabIndex="-1" role="dialog">
-        <div className="modal-dialog modal-kill">
-          <div className="modal-content">
-            <div className="modal-body">
-              <h4>Are you sure you want to kill this job?</h4>
-            </div>
-            <div className="modal-footer">
-              <button onClick={onHideModal} className="btn btn-default">Close</button>
-              <button onClick={killJob} className="btn btn-danger">Kill</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-class RelatedJobs extends React.Component {
-  render() {
-    const {hover, job, relatives} = this.props;
-
-    if (relatives.length < 2) return null;
-
-    let data = relatives.map((j) => ({
-      start: j.startTime,
-      finish: j.finishTime || new Date(),
-      job: j,
-    }));
-    data = _.sortBy(data, 'start');
-
-    const fmt = (d) => `${secondFormat(d.job.duration())} ${d.job.name}`;
-    const links = (d) => `#/job/${d.job.id}`;
-    const fs = (d) => {
-      if (d.job.id == job.id) {
-        return COLOUR_SELECTED;
-      } else if (hover && d.job.id == hover.id) {
-        return COLOUR_HOVER;
-      } else {
-        return COLOUR_MAP;
-      }
-    };
-
-    return (
-      <div>
-        <h4>Related Jobs</h4>
-        <Waterfall
-          data={data}
-          barHeight={30}
-          lineHeight={40}
-          width={1200}
-          textFormat={fmt}
-          fillStyle={fs}
-          linkFormat={links}
-        />
-      </div>
-    );
-  }
-}
-
-
-class TaskStats extends React.Component {
-  render() {
-    const tasks = this.props.tasks.filter((t) => !t.bogus);
-    const durations = sample(tasks.map((x) => x.duration()).sort(), 100, _.identity);
-    return (
-      <div>
-        <h4>{this.props.title} Timing</h4>
-        {durations.length ? <BoxPlot data={durations} tickFormat={secondFormat} /> : null}
-      </div>
-    );
-  }
-}
-
-function sample(arr, limit, comparator) {
-  // Sample arr down to size limit using comparator to take the largest value at
-  // each step. Works best if arr is already sorted.
-  if (arr.length <= limit) return arr;
-  const rv = [];
-  const sampleSize = arr.length / limit;
-  for (let i = 0; i < arr.length / sampleSize; i += 1) {
-    const vals = arr.slice(i * sampleSize, (i + 1) * sampleSize);
-    if (vals.length !== 0) {
-      rv.push(_.max(vals, comparator));
-    }
-  }
-  return rv;
-}
-
-
-class TaskWaterfall extends React.Component {
-  render() {
-    let {tasks} = this.props;
-    tasks.maps.forEach((d) => d.type = 'map');
-    tasks = tasks.maps.concat(tasks.reduces).filter((t) => !t.bogus).map((t) => {
-      return {start: t.startTime, finish: t.finishTime || new Date(), type: t.type};
-    });
-    const data = sample(_.sortBy(tasks, 'start'), 400, (d) => d.finish - d.start);
-    return (
-      <div>
-        <h4>Map <i className="map-label" /> / <i className="reduce-label" /> Reduce Waterfall</h4>
-        {data.length ? <Waterfall data={data} /> : null}
-      </div>
-    );
-  }
-}
-
-
-class Waterfall extends React.Component {
-  componentDidMount() {
-    waterfall(this.props.data, ReactDOM.findDOMNode(this), this.props);
-  }
-
-  shouldComponentUpdate() {
-    d3.select(ReactDOM.findDOMNode(this)).selectAll('svg').remove();
-    waterfall(this.props.data, ReactDOM.findDOMNode(this), this.props);
-    return false;
-  }
-
-  render() {
-    return <div />;
-  }
-}
-
-function waterfall(data, node, optsIn) {
-  const defaults = {
-    lineHeight: 1,
-    barHeight: 1,
-    width: 550,
-    textFormat: (t) => '',
-    linkFormat: null,
-    fillStyle: (d) => {
-      return d.type == 'map' ? COLOUR_MAP : COLOUR_REDUCE;
-    },
-  };
-  const opts = _.extend(defaults, optsIn);
-
-  const margin = {
-    top: 10, right: 20, bottom: 20, left: 20,
-  };
-  const width = opts.width - margin.left - margin.right;
-  const height = Math.max(100, data.length * opts.lineHeight) - margin.top - margin.bottom;
-
-  const chart = d3.waterfall()
-    .width(width)
-    .height(height)
-    .barHeight(opts.barHeight)
-    .textFormat(opts.textFormat)
-    .linkFormat(opts.linkFormat)
-    .barStyle(opts.fillStyle);
-
-  const start = d3.min(_.pluck(data, 'start'));
-  const finish = d3.max(_.pluck(data, 'finish'));
-  chart.domain([start, finish]);
-
-  if (((finish - start) / 1000) < 180) {
-    // Show seconds if the x domain is less than three minutes.
-    chart.tickFormat(d3.time.format.utc('%H:%M:%S'));
-  }
-
-  d3.select(node).selectAll('svg')
-    .data([data])
-    .enter().append('svg')
-    .attr('class', 'waterfall')
-    .attr('width', width + margin.left + margin.right)
-    .attr('height', height + margin.bottom + margin.top)
-    .append('g')
-    .attr('transform', `translate(${margin.left},${margin.top})`)
-    .call(chart);
-}
-
-
-class BoxPlot extends React.Component {
-  componentDidMount() {
-    boxplot(this.props.data, ReactDOM.findDOMNode(this), this.props.tickFormat);
-  }
-
-  shouldComponentUpdate() {
-    d3.select(ReactDOM.findDOMNode(this)).selectAll('svg').remove();
-    boxplot(this.props.data, ReactDOM.findDOMNode(this), this.props.tickFormat);
-    return false;
-  }
-
-  render() {
-    return <div />;
-  }
-}
-
-function boxplot(data, node, tickFormat) {
-  const margin = {
-    top: 10,
-    right: 100,
-    bottom: 20,
-    left: 100,
-  };
-  const width = 220 - margin.left - margin.right;
-  const height = 400 - margin.top - margin.bottom;
-
-  const chart = d3.box()
-    .whiskers(iqr(1.5))
-    .width(width)
-    .height(height)
-    .tickFormat(tickFormat);
-
-  chart.domain(d3.extent(data));
-
-  d3.select(node).selectAll('svg')
-    .data([data])
-    .enter().append('svg')
-    .attr('class', 'box')
-    .attr('width', width + margin.left + margin.right)
-    .attr('height', height + margin.bottom + margin.top)
-    .append('g')
-    .attr('transform', `translate(${margin.left},${margin.top})`)
-    .call(chart);
-
-  function iqr(k) {
-    return function(d) {
-      const q1 = d.quartiles[0];
-      const q3 = d.quartiles[2];
-      const iqrange = (q3 - q1) * k;
-      let i = -1;
-      let j = d.length;
-      while (d[i += 1] < q1 - iqrange);
-      while (d[j -= 1] > q3 + iqrange);
-      return [i, j];
-    };
-  }
-}
-
-
-class Summarizer extends React.Component {
-  render() {
-    const {progress} = this.props;
-    const inputRecords = this.props.input_records;
-    const outputRecords = this.props.output_records;
-    const recordsPerSec = inputRecords != notAvailable ? Math.floor(inputRecords / (progress.totalTime / 1000)) : null;
-    const computeTime = recordsPerSec ?
-      <span>{humanFormat(progress.totalTime)}<br />{numFormat(recordsPerSec)} {plural(recordsPerSec, 'record')}/sec</span>
-      : <span>{humanFormat(progress.totalTime)}</span>;
-    const pairs = [
-      ['Progress', <ProgressBar value={progress.progress} />],
-      ['Total', numFormat(progress.total)],
-      ['Completed', numFormat(progress.completed)],
-      ['Running', numFormat(progress.running)],
-      ['Pending', numFormat(progress.pending)],
-      ['Killed', numFormat(progress.killed)],
-      ['Failed', numFormat(progress.failed)],
-      ['Input Records', inputRecords != notAvailable ? numFormat(inputRecords) : null],
-      ['Output Records', outputRecords != notAvailable ? numFormat(outputRecords) : null],
-      ['Compute Time', progress.totalTime > 0 ? computeTime : null],
-    ];
-    return (
-      <table className="table">
-        <tbody>{pairs.map((t, i) => <tr key={i}><th>{t[0]}</th><td>{t[1]}</td></tr>)}</tbody>
-      </table>
-    );
-  }
-}
-
-
-class MapSummary extends React.Component {
-  render() {
-    const {job} = this.props;
-    const inputRecords = this.props.counters.get('TaskCounter.MAP_INPUT_RECORDS');
-    const outputRecords = this.props.counters.get('TaskCounter.MAP_OUTPUT_RECORDS');
-    return <Summarizer progress={job.maps} input_records={inputRecords.map} output_records={outputRecords.map} />;
-  }
-}
-
-
-class ReduceSummary extends React.Component {
-  render() {
-    const {job} = this.props;
-    const inputRecords = this.props.counters.get('TaskCounter.REDUCE_INPUT_RECORDS');
-    const outputRecords = this.props.counters.get('TaskCounter.REDUCE_OUTPUT_RECORDS');
-    return <Summarizer progress={job.reduces} input_records={inputRecords.reduce} output_records={outputRecords.reduce} />;
   }
 }
